# docs/openapi/roomfit-api.yaml
openapi: 3.0.3
info:
  title: RoomFit AI Backend API
  version: 1.0.0
  description: |
    RoomFit AI MVP backend API specification.

    RoomFit MVP Demo Flow:
    1. GET /api/rooms/1
    2. GET /api/products/mock
    3. PUT /api/rooms/1/furniture
    4. POST /api/agent/context
    5. POST /api/layouts/recommend
    6. POST /api/layouts/validate
    7. PUT /api/layouts/{layoutId}

    Frontend integration notes:
    - 단위는 meter입니다.
    - 좌표계는 x-z 평면입니다.
    - position은 가구 중심 좌표입니다.
    - rotation은 degree입니다.
    - DELETED 가구는 렌더링/검증에서 제외하는 것이 자연스럽습니다.
    - POST /api/layouts/validate는 저장하지 않습니다.
    - PUT /api/layouts/{layoutId}는 저장 + 재검증 + 점수 재계산을 수행합니다.
    - 현재 MVP는 in-memory repository를 사용하므로 서버 재시작 시 업로드 room/context/layout 데이터가 초기화됩니다.
servers:
  - url: http://localhost:8080
tags:
  - name: Rooms
    description: 방 조회, RoomPlan 업로드, 기존 가구 상태 수정 API
  - name: Styles
    description: 스타일 이미지 선택용 API
  - name: Products
    description: MVP 제품 카드 렌더링용 Mock Product API
  - name: Agent
    description: 사용자 목표/스타일/선택 제품 기반 추천 Context 생성 API
  - name: Layouts
    description: 추천 배치 생성, 드래그 검증, 수정 저장, 피드백, 확정 API
paths:
  /api/rooms/samples:
    get:
      tags: [Rooms]
      summary: 샘플 방 목록 조회
      description: 데모 시작 시 사용할 샘플 방 목록을 조회합니다.
      responses:
        "200":
          description: Sample room list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomListResponse"
  /api/rooms/upload:
    post:
      tags: [Rooms]
      summary: RoomPlan JSON 업로드
      description: iOS RoomPlan App 또는 Mock/Manual 입력에서 생성한 RoomFit JSON을 백엔드에 저장하고 roomId를 발급합니다. 업로드된 방은 GET /api/rooms/{roomId}로 조회할 수 있습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomUploadRequest"
            example:
              name: RoomPlan Scan Room
              room:
                width: 3.2
                depth: 4.5
                height: 2.4
                unit: meter
              openings:
                - id: door-1
                  type: door
                  wall: south
                  offset: 0.7
                  width: 0.8
                  height: 2.1
                - id: window-1
                  type: window
                  wall: north
                  offset: 1.5
                  width: 1.2
                  height: 1.0
                  sillHeight: 0.9
              furniture:
                - id: bed-1
                  type: bed
                  label: 침대
                  width: 1.1
                  depth: 2.0
                  height: 0.45
                  position:
                    x: 0.8
                    z: 1.4
                  rotation: 0
                  status: EXISTING
      responses:
        "201":
          description: Uploaded room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomCommonResponse"
        "400":
          $ref: "#/components/responses/InvalidRoomUpload"
  /api/rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: 특정 방 조회
      description: roomId 기준으로 방 크기, 문/창문, 기존 가구 목록을 조회합니다. 프론트는 이 응답을 기준으로 방을 렌더링합니다.
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomCommonResponse"
        "404":
          $ref: "#/components/responses/RoomNotFound"
  /api/rooms/{roomId}/furniture:
    put:
      tags: [Rooms]
      summary: 기존 가구 상태/위치 수정
      description: 기존 가구를 삭제하거나 위치/회전/상태를 수정합니다. 예를 들어 기존 책상을 DELETED로 변경한 뒤 새 책상을 추천받을 수 있습니다.
      parameters:
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FurnitureUpdateRequest"
            example:
              furnitureUpdates:
                - id: desk-1
                  status: DELETED
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomCommonResponse"
        "400":
          $ref: "#/components/responses/InvalidFurnitureStatus"
        "404":
          $ref: "#/components/responses/FurnitureNotFound"
  /api/styles/images:
    get:
      tags: [Styles]
      summary: 스타일 이미지 목록 조회
      description: 사용자가 선호 스타일을 선택할 수 있도록 스타일 이미지 목록을 반환합니다.
      responses:
        "200":
          description: Style images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleImageListResponse"
  /api/products/mock:
    get:
      tags: [Products]
      summary: Mock 제품 목록 조회
      description: MVP 데모에서 사용자가 선택할 수 있는 제품 카드 목록을 반환합니다. selectedProductIds로 Agent Context 생성 시 사용됩니다.
      responses:
        "200":
          description: Mock products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MockProductListResponse"
  /api/agent/context:
    post:
      tags: [Agent]
      summary: Agent Context 생성
      description: 방 ID, 생활 목표, 디자인 스타일, 필수/선택 가구, 선택 제품 정보를 바탕으로 추천에 사용할 Context를 생성합니다. 이후 POST /api/layouts/recommend에 contextId를 전달합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentContextRequest"
            example:
              roomId: 1
              lifestyleGoal: STUDY_FOCUSED
              designStyle: [MINIMAL, WHITE_TONE]
              requiredItems: [bed, desk, chair]
              optionalItems: [storage, rug, lamp]
              selectedImageIds: [1, 3]
              selectedProductIds: [desk-01, chair-01, lamp-01]
      responses:
        "201":
          description: Created context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentContextCommonResponse"
        "400":
          description: Invalid request body or selected product id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorCommonResponse"
        "404":
          $ref: "#/components/responses/RoomNotFound"
  /api/layouts/recommend:
    post:
      tags: [Layouts]
      summary: 배치 추천 생성
      description: Agent Context를 기반으로 기존 가구와 선택 제품을 고려한 추천 배치를 생성합니다. recommendedFurniture, validationResult, scoreSummary를 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
            example:
              contextId: 1
      responses:
        "201":
          description: Recommended layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutCommonResponse"
        "404":
          $ref: "#/components/responses/ContextNotFound"
  /api/layouts/validate:
    post:
      tags: [Layouts]
      summary: 현재 배치 검증
      description: 사용자가 가구를 드래그하는 중 현재 화면의 가구 배치가 충돌/경계/문/창문/동선 조건을 만족하는지 검증합니다. 저장은 수행하지 않습니다. furniture 배열에는 현재 layoutId에 포함된 전체 가구 목록을 전달해야 합니다. 일부 가구만 전달하면 FURNITURE_ARRAY_MISMATCH가 발생할 수 있습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateRequest"
            example:
              layoutId: 1
              furniture:
                - id: bed-1
                  position: { x: 0.8, z: 1.4 }
                  rotation: 0
                  status: EXISTING
                - id: desk-rec-1
                  position: { x: 2.3, z: 1.0 }
                  rotation: 0
                  status: USER_MODIFIED
                - id: chair-rec-1
                  position: { x: 2.3, z: 1.8 }
                  rotation: 0
                  status: USER_MODIFIED
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationCommonResponse"
        "400":
          $ref: "#/components/responses/FurnitureArrayMismatch"
        "404":
          $ref: "#/components/responses/LayoutNotFound"
  /api/layouts/{layoutId}:
    put:
      tags: [Layouts]
      summary: 수정 배치 저장
      description: 사용자가 수정한 최종 가구 배치를 저장하고 validationResult와 scoreSummary를 다시 계산합니다.
      parameters:
        - $ref: "#/components/parameters/LayoutId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LayoutUpdateRequest"
            example:
              furniture:
                - id: bed-1
                  position: { x: 0.8, z: 1.4 }
                  rotation: 0
                  status: EXISTING
                - id: desk-rec-1
                  position: { x: 2.3, z: 1.0 }
                  rotation: 0
                  status: USER_MODIFIED
                - id: chair-rec-1
                  position: { x: 2.3, z: 1.8 }
                  rotation: 0
                  status: USER_MODIFIED
      responses:
        "200":
          description: Updated layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LayoutCommonResponse"
        "400":
          $ref: "#/components/responses/FurnitureArrayMismatch"
        "404":
          $ref: "#/components/responses/LayoutNotFound"
        "409":
          $ref: "#/components/responses/AlreadyConfirmed"
  /api/layouts/feedback:
    post:
      tags: [Layouts]
      summary: 사용자 피드백 기반 재추천
      description: 사용자의 자연어 피드백을 바탕으로 기존 배치를 조정하거나 재추천합니다. MVP에서는 제한적인 규칙 기반 피드백을 지원합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
            examples:
              largerDesk:
                value:
                  layoutId: 1
                  feedback: 책상 더 크게
              moreStorage:
                value:
                  layoutId: 1
                  feedback: 수납 늘려줘
              openSpace:
                value:
                  layoutId: 1
                  feedback: 방이 넓어 보이게
      responses:
        "201":
          description: Feedback recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackCommonResponse"
        "400":
          $ref: "#/components/responses/UnsupportedFeedbackIntent"
        "404":
          $ref: "#/components/responses/LayoutNotFound"
  /api/layouts/{layoutId}/confirm:
    post:
      tags: [Layouts]
      summary: 최종 배치 확정
      description: 사용자가 추천 또는 수정한 배치를 최종 확정합니다.
      parameters:
        - $ref: "#/components/parameters/LayoutId"
      responses:
        "200":
          description: Confirmed layout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmCommonResponse"
        "404":
          $ref: "#/components/responses/LayoutNotFound"
        "409":
          $ref: "#/components/responses/AlreadyConfirmed"
components:
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    LayoutId:
      name: layoutId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    ContextNotFound:
      description: Context not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    LayoutNotFound:
      description: Layout not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    FurnitureNotFound:
      description: Furniture not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    InvalidFurnitureStatus:
      description: Invalid furniture status
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    FurnitureArrayMismatch:
      description: Furniture array mismatch
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    UnsupportedFeedbackIntent:
      description: Unsupported feedback intent
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    AlreadyConfirmed:
      description: Already confirmed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    InvalidRequestBody:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
    InvalidRoomUpload:
      description: Invalid uploaded room data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorCommonResponse"
  schemas:
    CommonResponse:
      type: object
      required: [success, data, error]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      properties:
        code:
          type: string
          enum:
            - ROOM_NOT_FOUND
            - CONTEXT_NOT_FOUND
            - LAYOUT_NOT_FOUND
            - FURNITURE_NOT_FOUND
            - STYLE_IMAGE_NOT_FOUND
            - PRODUCT_NOT_FOUND
            - INVALID_FURNITURE_STATUS
            - INVALID_LIFESTYLE_GOAL
            - INVALID_DESIGN_STYLE
            - INVALID_FURNITURE_TYPE
            - INVALID_ROOM_DIMENSION
            - INVALID_OPENING_DATA
            - REQUIRED_ITEM_EMPTY
            - STYLE_IMAGE_EMPTY
            - INVALID_FURNITURE_POSITION
            - INVALID_ROTATION
            - FURNITURE_ARRAY_MISMATCH
            - UNSUPPORTED_FEEDBACK_INTENT
            - INVALID_REQUEST_BODY
            - ALREADY_CONFIRMED
            - RECOMMENDATION_FAILED
          example: ROOM_NOT_FOUND
        message:
          type: string
    ErrorCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            success:
              example: false
            data:
              nullable: true
              example: null
            error:
              $ref: "#/components/schemas/ErrorPayload"
    RoomListResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/RoomResponse"
    RoomCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/RoomResponse"
    RoomResponse:
      type: object
      description: 방 조회/업로드 응답. 프론트는 room/openings/furniture를 기준으로 방을 렌더링합니다.
      properties:
        roomId:
          type: integer
          format: int64
          description: 방 ID
          example: 1
        name:
          type: string
          description: 방 이름
          example: RoomPlan Scan Room
        room:
          description: 방 크기 정보. 모든 단위는 meter입니다.
          $ref: "#/components/schemas/RoomDimension"
        openings:
          type: array
          items:
            $ref: "#/components/schemas/Opening"
        furniture:
          type: array
          items:
            $ref: "#/components/schemas/Furniture"
        source:
          type: string
          enum: [SAMPLE, ROOMPLAN]
        createdAt:
          type: string
          format: date-time
    RoomUploadRequest:
      type: object
      description: RoomPlan iOS App 또는 Mock/Manual 입력에서 생성한 RoomFit JSON 업로드 요청
      required: [room]
      properties:
        name:
          type: string
          default: Uploaded Room
        room:
          $ref: "#/components/schemas/RoomDimension"
        openings:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/Opening"
        furniture:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/Furniture"
    RoomDimension:
      type: object
      description: 방 크기. width/depth/height는 meter 단위입니다.
      properties:
        width:
          type: number
          format: double
          description: 방 가로 길이(meter)
          example: 3.2
        depth:
          type: number
          format: double
          description: 방 깊이(meter)
          example: 4.5
        height:
          type: number
          format: double
          description: 방 높이(meter)
          example: 2.4
        unit:
          type: string
          example: meter
    Opening:
      type: object
      description: 방의 문/창문 정보
      properties:
        id:
          type: string
        type:
          type: string
          enum: [door, window]
        wall:
          type: string
          enum: [north, south, east, west]
        offset:
          type: number
        width:
          type: number
        height:
          type: number
        sillHeight:
          type: number
          nullable: true
    Furniture:
      type: object
      description: 방 안의 가구. width/depth/height는 meter, position은 x-z 평면 중심 좌표입니다.
      properties:
        id:
          type: string
        type:
          type: string
          enum: [bed, desk, chair, storage, rug, lamp]
        label:
          type: string
        width:
          type: number
          description: 가구 가로 길이(meter)
        depth:
          type: number
          description: 가구 깊이(meter)
        height:
          type: number
          description: 가구 높이(meter)
        position:
          description: x-z 평면에서의 가구 중심 좌표
          $ref: "#/components/schemas/Position"
        rotation:
          type: number
          description: degree 단위 회전 각도
        status:
          type: string
          description: 가구 상태. DELETED는 렌더링/검증에서 제외하는 것이 자연스럽습니다.
          enum: [EXISTING, DELETED, RECOMMENDED, USER_MODIFIED]
        productId:
          type: string
          description: 선택 제품 기반 추천인 경우 제품 ID. 기존 가구는 null 허용
          nullable: true
        styleTags:
          description: 추천/스타일 계산에 사용되는 태그
          type: array
          items:
            type: string
    Position:
      type: object
      description: x-z 평면에서의 중심 좌표
      properties:
        x:
          type: number
          description: x축 중심 좌표(meter)
          example: 2.3
        z:
          type: number
          description: z축 중심 좌표(meter)
          example: 1.0
    FurnitureUpdateRequest:
      type: object
      properties:
        furnitureUpdates:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              status:
                type: string
                enum: [EXISTING, DELETED, RECOMMENDED, USER_MODIFIED]
    StyleImageListResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/StyleImage"
    StyleImage:
      type: object
      properties:
        imageId:
          type: integer
          format: int64
        title:
          type: string
        imageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
    MockProductListResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/MockProduct"
    MockProduct:
      type: object
      description: MVP 제품 카드 렌더링용 Mock Product 응답
      properties:
        productId:
          type: string
          description: 제품 ID. Agent Context의 selectedProductIds에 사용됩니다.
          example: desk-01
        type:
          type: string
          enum: [bed, desk, chair, storage, rug, lamp]
        name:
          type: string
        brand:
          type: string
        width:
          type: number
        depth:
          type: number
        height:
          type: number
        price:
          type: integer
        styleTags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        requiredClearance:
          $ref: "#/components/schemas/RequiredClearance"
    RequiredClearance:
      type: object
      properties:
        front:
          type: number
        side:
          type: number
    AgentContextRequest:
      type: object
      description: 사용자 목표/스타일/선택 제품 기반 추천 Context 생성 요청
      properties:
        roomId:
          type: integer
          format: int64
          description: 추천 대상 방 ID
          example: 1
        lifestyleGoal:
          type: string
          enum: [STUDY_FOCUSED, RELAX_FOCUSED, STORAGE_FOCUSED, WFH_FOCUSED]
        designStyle:
          type: array
          items:
            type: string
            enum: [MINIMAL, NATURAL, WHITE_TONE, WOOD_TONE, COZY, MODERN]
        requiredItems:
          type: array
          items:
            type: string
            enum: [bed, desk, chair, storage, rug, lamp]
        optionalItems:
          type: array
          items:
            type: string
            enum: [bed, desk, chair, storage, rug, lamp]
        selectedImageIds:
          type: array
          items:
            type: integer
            format: int64
        selectedProductIds:
          description: 선택한 Mock Product ID 목록. 추천 가구 크기/productId/styleTags에 반영됩니다.
          type: array
          items:
            type: string
    AgentContextCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AgentContextResponse"
    AgentContextResponse:
      type: object
      description: 추천에 사용할 Agent Context 응답
      properties:
        contextId:
          type: integer
          format: int64
          description: 추천 Context ID
          example: 1
        roomId:
          type: integer
          format: int64
        lifestyleGoal:
          type: string
        designStyle:
          type: array
          items:
            type: string
        requiredItems:
          type: array
          items:
            type: string
        optionalItems:
          type: array
          items:
            type: string
        selectedImageIds:
          type: array
          items:
            type: integer
            format: int64
        selectedProductIds:
          type: array
          items:
            type: string
        styleTags:
          type: array
          items:
            type: string
        selectedProducts:
          type: array
          items:
            $ref: "#/components/schemas/MockProduct"
        createdAt:
          type: string
          format: date-time
    RecommendRequest:
      type: object
      description: 배치 추천 생성 요청
      properties:
        contextId:
          type: integer
          format: int64
          description: 추천에 사용할 Agent Context ID
          example: 1
    LayoutCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/LayoutResponse"
    LayoutResponse:
      type: object
      description: 추천/수정 저장 배치 응답. 프론트는 recommendedFurniture를 Three.js 가구 렌더링에 사용합니다.
      properties:
        layoutId:
          type: integer
          format: int64
          description: 배치 ID
          example: 1
        status:
          type: string
          nullable: true
          enum: [SUCCESS, FALLBACK]
        recommendedFurniture:
          type: array
          items:
            $ref: "#/components/schemas/Furniture"
        scoreSummary:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ScoreSummary"
        validationResult:
          $ref: "#/components/schemas/ValidationResult"
    ScoreSummary:
      type: object
      description: 추천 점수 요약. validationResult, 생활 목표, 스타일 태그 일치도를 기반으로 계산됩니다.
      properties:
        totalScore:
          type: integer
        collisionScore:
          type: integer
        boundaryScore:
          type: integer
        doorWindowScore:
          type: integer
        pathScore:
          type: integer
        goalScore:
          type: integer
        styleScore:
          type: integer
    ValidationResult:
      type: object
      description: 배치 검증 결과. 프론트 체크리스트와 경고 표시용으로 사용합니다.
      properties:
        collisionFree:
          type: boolean
          description: 가구 간 충돌이 없는지 여부
        boundaryValid:
          type: boolean
        doorClearance:
          type: boolean
        windowClearance:
          type: boolean
        pathSecured:
          type: boolean
        validationItems:
          description: 프론트 체크리스트 UI용 검증 항목
          type: array
          items:
            $ref: "#/components/schemas/ValidationItem"
        warnings:
          type: array
          items:
            type: string
    ValidationItem:
      type: object
      description: 프론트 체크리스트 UI용 단일 검증 항목
      properties:
        type:
          type: string
          description: 검증 타입
          enum: [collision, boundary, door_clearance, window_clearance, path]
        passed:
          type: boolean
        message:
          type: string
    FurniturePositionDto:
      type: object
      description: 검증/수정 저장 시 사용하는 가구 위치 DTO
      properties:
        id:
          type: string
          description: 가구 ID. 기존 recommendedFurniture의 id와 일치해야 합니다.
        position:
          $ref: "#/components/schemas/Position"
        rotation:
          type: number
        status:
          type: string
          nullable: true
          enum: [EXISTING, DELETED, RECOMMENDED, USER_MODIFIED]
    ValidateRequest:
      type: object
      description: 현재 화면 배치 검증 요청. 저장은 수행하지 않습니다. furniture 배열에는 현재 layoutId에 포함된 전체 가구 목록을 전달해야 합니다. 일부 가구만 전달하면 FURNITURE_ARRAY_MISMATCH가 발생할 수 있습니다.
      properties:
        layoutId:
          type: integer
          format: int64
        furniture:
          description: 프론트 화면의 현재 가구 배치 전체 배열. 현재 layoutId에 포함된 전체 가구 목록을 전달해야 하며, 일부 가구만 전달하면 FURNITURE_ARRAY_MISMATCH가 발생할 수 있습니다.
          type: array
          items:
            $ref: "#/components/schemas/FurniturePositionDto"
    LayoutUpdateRequest:
      type: object
      description: 수정 배치 저장 요청. 저장 후 validationResult와 scoreSummary를 다시 계산합니다.
      properties:
        furniture:
          type: array
          items:
            $ref: "#/components/schemas/FurniturePositionDto"
    ValidationCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ValidationResult"
    FeedbackRequest:
      type: object
      description: 사용자 자연어 피드백 기반 재추천 요청
      properties:
        layoutId:
          type: integer
          format: int64
        feedback:
          type: string
          enum: [책상 더 크게, 수납 늘려줘, 방이 넓어 보이게]
    FeedbackCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/FeedbackResponse"
    FeedbackResponse:
      type: object
      description: 사용자 피드백 기반 재추천 응답
      properties:
        layoutId:
          type: integer
          format: int64
        status:
          type: string
          enum: [SUCCESS, FALLBACK]
        recommendedFurniture:
          type: array
          items:
            $ref: "#/components/schemas/Furniture"
        scoreSummary:
          $ref: "#/components/schemas/ScoreSummary"
        validationResult:
          $ref: "#/components/schemas/ValidationResult"
        interpretedIntent:
          type: object
          additionalProperties: true
    ConfirmCommonResponse:
      allOf:
        - $ref: "#/components/schemas/CommonResponse"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ConfirmResponse"
    ConfirmResponse:
      type: object
      properties:
        layoutId:
          type: integer
          format: int64
        confirmed:
          type: boolean
        confirmedAt:
          type: string
          format: date-time